Move code to instructor_task Django app.
