+++ lms/djangoapps/instructor_task/tasks_helper.py
@@ -11,7 +11,7 @@ from time import time
 from sys import exc_info
 from traceback import format_exc
 
-from celery import task, current_task
+from celery import current_task
 from celery.utils.log import get_task_logger
 from celery.states import SUCCESS, FAILURE
 
@@ -24,10 +24,10 @@ from xmodule.modulestore.django import modulestore
 import mitxmako.middleware as middleware
 from track.views import task_track
 
-from courseware.models import StudentModule, CourseTask
+from courseware.models import StudentModule
 from courseware.model_data import ModelDataCache
 from courseware.module_render import get_module_for_descriptor_internal
-
+from instructor_task.models import InstructorTask
 
 # define different loggers for use within tasks and on client side
 TASK_LOG = get_task_logger(__name__)
@@ -78,7 +78,7 @@ def _perform_module_state_update(course_id, module_state_key, student_identifier
           'duration_ms': how long the task has (or had) been running.
 
     Because this is run internal to a task, it does not catch exceptions.  These are allowed to pass up to the
-    next level, so that it can set the failure modes and capture the error trace in the CourseTask and the
+    next level, so that it can set the failure modes and capture the error trace in the InstructorTask and the
     result object.
 
     """
@@ -157,7 +157,7 @@ def _perform_module_state_update(course_id, module_state_key, student_identifier
 
 @transaction.autocommit
 def _save_course_task(course_task):
-    """Writes CourseTask course_task immediately, ensuring the transaction is committed."""
+    """Writes InstructorTask course_task immediately, ensuring the transaction is committed."""
     course_task.save()
 
 
@@ -166,7 +166,7 @@ def _update_problem_module_state(entry_id, course_id, module_state_key, student_
     """
     Performs generic update by visiting StudentModule instances with the update_fcn provided.
 
-    The `entry_id` is the primary key for the CourseTask entry representing the task.  This function
+    The `entry_id` is the primary key for the InstructorTask entry representing the task.  This function
     updates the entry on success and failure of the _perform_module_state_update function it
     wraps.  It is setting the entry's value for task_state based on what Celery would set it to once
     the task returns to Celery:  FAILURE if an exception is encountered, and SUCCESS if it returns normally.
@@ -181,9 +181,9 @@ def _update_problem_module_state(entry_id, course_id, module_state_key, student_
               Pass-through of input `action_name`.
           'duration_ms': how long the task has (or had) been running.
 
-    Before returning, this is also JSON-serialized and stored in the task_output column of the CourseTask entry.
+    Before returning, this is also JSON-serialized and stored in the task_output column of the InstructorTask entry.
 
-    If exceptions were raised internally, they are caught and recorded in the CourseTask entry.
+    If exceptions were raised internally, they are caught and recorded in the InstructorTask entry.
     This is also a JSON-serialized dict, stored in the task_output column, containing the following keys:
 
            'exception':  type of exception object
@@ -199,9 +199,9 @@ def _update_problem_module_state(entry_id, course_id, module_state_key, student_
     fmt = 'Starting to update problem modules as task "{task_id}": course "{course_id}" problem "{state_key}": nothing {action} yet'
     TASK_LOG.info(fmt.format(task_id=task_id, course_id=course_id, state_key=module_state_key, action=action_name))
 
-    # get the CourseTask to be updated.  If this fails, then let the exception return to Celery.
+    # get the InstructorTask to be updated.  If this fails, then let the exception return to Celery.
     # There's no point in catching it here.
-    entry = CourseTask.objects.get(pk=entry_id)
+    entry = InstructorTask.objects.get(pk=entry_id)
     entry.task_id = task_id
     _save_course_task(entry)
 
@@ -228,7 +228,7 @@ def _update_problem_module_state(entry_id, course_id, module_state_key, student_
         _save_course_task(entry)
         raise
 
-    # if we get here, we assume we've succeeded, so update the CourseTask entry in anticipation:
+    # if we get here, we assume we've succeeded, so update the InstructorTask entry in anticipation:
     entry.task_output = json.dumps(task_progress)
     entry.task_state = SUCCESS
     _save_course_task(entry)
@@ -329,39 +329,6 @@ def _rescore_problem_module_state(module_descriptor, student_module, xmodule_ins
         return True
 
 
-def _filter_module_state_for_done(modules_to_update):
-    """Filter to apply for rescoring, to limit module instances to those marked as done"""
-    return modules_to_update.filter(state__contains='"done": true')
-
-
-@task
-def rescore_problem(entry_id, course_id, task_input, xmodule_instance_args):
-    """Rescores problem in `course_id`.
-
-    `entry_id` is the id value of the CourseTask entry that corresponds to this task.
-    `course_id` identifies the course.
-    `task_input` should be a dict with the following entries:
-
-      'problem_url': the full URL to the problem to be rescored.  (required)
-      'student': the identifier (username or email) of a particular user whose
-          problem submission should be rescored.  If not specified, all problem
-          submissions will be rescored.
-
-    `xmodule_instance_args` provides information needed by _get_module_instance_for_task()
-    to instantiate an xmodule instance.
-    """
-    action_name = 'rescored'
-    update_fcn = _rescore_problem_module_state
-    filter_fcn = lambda(modules_to_update): modules_to_update.filter(state__contains='"done": true')
-    problem_url = task_input.get('problem_url')
-    student_ident = None
-    if 'student' in task_input:
-        student_ident = task_input['student']
-    return _update_problem_module_state(entry_id, course_id, problem_url, student_ident,
-                                        update_fcn, action_name, filter_fcn=filter_fcn,
-                                        xmodule_instance_args=xmodule_instance_args)
-
-
 @transaction.autocommit
 def _reset_problem_attempts_module_state(_module_descriptor, student_module, xmodule_instance_args=None):
     """
@@ -388,27 +355,6 @@ def _reset_problem_attempts_module_state(_module_descriptor, student_module, xmo
     return True
 
 
-@task
-def reset_problem_attempts(entry_id, course_id, task_input, xmodule_instance_args):
-    """Resets problem attempts to zero for `problem_url` in `course_id` for all students.
-
-    `entry_id` is the id value of the CourseTask entry that corresponds to this task.
-    `course_id` identifies the course.
-    `task_input` should be a dict with the following entries:
-
-      'problem_url': the full URL to the problem to be rescored.  (required)
-
-    `xmodule_instance_args` provides information needed by _get_module_instance_for_task()
-    to instantiate an xmodule instance.
-    """
-    action_name = 'reset'
-    update_fcn = _reset_problem_attempts_module_state
-    problem_url = task_input.get('problem_url')
-    return _update_problem_module_state(entry_id, course_id, problem_url, None,
-                                        update_fcn, action_name, filter_fcn=None,
-                                        xmodule_instance_args=xmodule_instance_args)
-
-
 @transaction.autocommit
 def _delete_problem_module_state(_module_descriptor, student_module, xmodule_instance_args=None):
     """
@@ -423,24 +369,3 @@ def _delete_problem_module_state(_module_descriptor, student_module, xmodule_ins
     task_info = {"student": student_module.student.username, "task_id": _get_task_id_from_xmodule_args(xmodule_instance_args)}
     task_track(request_info, task_info, 'problem_delete_state', {}, page='x_module_task')
     return True
-
-
-@task
-def delete_problem_state(entry_id, course_id, task_input, xmodule_instance_args):
-    """Deletes problem state entirely for `problem_url` in `course_id` for all students.
-
-    `entry_id` is the id value of the CourseTask entry that corresponds to this task.
-    `course_id` identifies the course.
-    `task_input` should be a dict with the following entries:
-
-      'problem_url': the full URL to the problem to be rescored.  (required)
-
-    `xmodule_instance_args` provides information needed by _get_module_instance_for_task()
-    to instantiate an xmodule instance.
-    """
-    action_name = 'deleted'
-    update_fcn = _delete_problem_module_state
-    problem_url = task_input.get('problem_url')
-    return _update_problem_module_state(entry_id, course_id, problem_url, None,
-                                        update_fcn, action_name, filter_fcn=None,
-                                        xmodule_instance_args=xmodule_instance_args)
