+++ lms/envs/common.py
@@ -124,8 +124,8 @@ MITX_FEATURES = {
     # Do autoplay videos for students
     'AUTOPLAY_VIDEOS': True,
 
-    # Enable instructor dash to submit course-level background tasks
-    'ENABLE_COURSE_BACKGROUND_TASKS': True,
+    # Enable instructor dash to submit background tasks
+    'ENABLE_INSTRUCTOR_BACKGROUND_TASKS': True,
 }
 
 # Used for A/B testing
@@ -694,6 +694,7 @@ INSTALLED_APPS = (
     'util',
     'certificates',
     'instructor',
+    'instructor_task',
     'open_ended_grading',
     'psychometrics',
     'licenses',
